% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affirmations.R
\name{create_affirmation}
\alias{create_affirmation}
\title{Create a personalized affirmation}
\usage{
create_affirmation(name, personality)
}
\arguments{
\item{name}{A string with the user's name.}

\item{personality}{A string like "Softie" or "Brat Summer Girl".}
}
\description{
Combines a user’s name and personality to generate a personalized affirmation.
}
\details{
Returns a cheerful affirmation string.
}
\examples{
create_affirmation("Alex", "Softie")
}
