% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affirmations.R
\name{give_affirmation}
\alias{give_affirmation}
\title{Give a general affirmation}
\usage{
give_affirmation()
}
\value{
A character string is printed to the console.
}
\description{
Prints a fixed, positive affirmation message.
}
\examples{
give_affirmation()

}
