% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_affirmation.R
\name{get_affirmation}
\alias{get_affirmation}
\title{Launch the Affirmation Shiny App}
\usage{
get_affirmation()
}
\description{
This function launches the full Shiny app included in the affirmations package.
}
